Add two animated morph target samples
Adds two new sample files that demonstrate morph targets and the
animation of those targets.

The first sample is a simple cube with two morph targets:
  * `thin` moves four vertices to create a thin cube
  * `angle` moves two vertices to create a ramp

The second sample is a simple sphere with two morph targets that are
more complicated:
  * `blob` moves many vertices to create an amorphous blob
  * `ship` moves even more vertices to create something that resembles
    a spaceship.

In both cases, there is an animation that animates between the two
morph targets attached to the object.

The [Babylon.js sandbox](https://sandbox.babylonjs.com/) is a great
way to try these models out.

Thanks to @BeardedGnome and @deltakosh for the
[Babylon.js](https://github.com/BabylonJS/Babylon.js) changes needed
to both support morph targets in general, as well as to the glTF-specific
loader.

Thanks also to @bghgary for the
[Unity Exporter](https://github.com/bghgary/glTF-Tools-for-Unity) help.